Module os_str_bytes::raw

Functions that cannot be implemented outside of this crate.

Due to the limited specification of the encodings used by this crate, some functions cannot be implemented compatibly. As a solution, this module contains definitions that will work for all supported platforms, by taking advantage of internal assumptions.

These functions should only be passed bytes that can be given to OsStrBytes::from_bytes without error. Only valid UTF-8 data or bytes extracted using this crate are acceptable. Other sequences will not cause safety issues, but they may result in panics or confusing results, so their use is unsupported.

Functions

ends_with
Returns true if and only if the encoded bytes end with the given suffix.

starts_with
Returns true if and only if the encoded bytes start with the given prefix.