Struct os_str_bytes::RawOsString

pub struct RawOsString(/* private fields */);

A container for owned byte strings converted by this crate.

For more information, see RawOsStr.

Implementations

impl RawOsString

pub fn new<S>(string: S) -> Self
where S: Into<OsString>,

Wraps a platform-native string, without copying or encoding conversion.

Examples

use std::env;

use os_str_bytes::RawOsString;

let os_string = env::current_exe()?.into_os_string();
println!("{:?}", RawOsString::new(os_string));

pub fn from_string(string: String) -> Self

๐Deprecated since 7.0.0: use new instead

Wraps a string, without copying or encoding conversion.

Examples

use os_str_bytes::RawOsString;

let string = "foobar".to_owned();
let raw = RawOsString::from_string(string.clone());
assert_eq!(string, raw);

pub unsafe fn from_encoded_vec_unchecked(string: Vec<u8>) -> Self

Equivalent to OsString::from_encoded_bytes_unchecked.

Examples

use std::env;

use os_str_bytes::RawOsString;

let os_string = env::current_exe()?.into_os_string();
let raw = RawOsString::new(os_string);
let raw_bytes = raw.clone().into_encoded_vec();
assert_eq!(raw, unsafe {
    RawOsString::from_encoded_vec_unchecked(raw_bytes)
});

pub fn clear(&mut self)

Equivalent to String::clear.

Examples

use std::env;

use os_str_bytes::RawOsString;

let os_string = env::current_exe()?.into_os_string();
let mut raw = RawOsString::new(os_string);
raw.clear();
assert!(raw.is_empty());

pub fn into_box(self) -> Box<RawOsStr>

Equivalent to String::into_boxed_str.

Examples

use os_str_bytes::RawOsString;

let string = "foobar".to_owned();
let raw = RawOsString::new(string.clone());
assert_eq!(string, *raw.into_box());

pub fn into_encoded_vec(self) -> Vec<u8>

Equivalent to OsString::into_encoded_bytes.

The returned string will not use the unspecified encoding. It can only be passed to methods accepting the internal encoding of OsStr, such as from_encoded_vec_unchecked.

Examples

use os_str_bytes::RawOsString;

let string = "foobar".to_owned();
let raw = RawOsString::new(string.clone());
assert_eq!(string.into_bytes(), raw.into_encoded_vec());

pub fn into_os_string(self) -> OsString

Converts this representation back to a platform-native string, without copying or encoding conversion.

Examples

use std::env;

use os_str_bytes::RawOsString;

let os_string = env::current_exe()?.into_os_string();
let raw = RawOsString::new(os_string.clone());
assert_eq!(os_string, raw.into_os_string());

pub fn into_string(self) -> Result<String, Self>

Equivalent to OsString::into_string.

Examples

use os_str_bytes::RawOsString;

let string = "foobar".to_owned();
let raw = RawOsString::new(string.clone());
assert_eq!(Ok(string), raw.into_string());

pub fn shrink_to_fit(&mut self)

Equivalent to String::shrink_to_fit.

Examples

use os_str_bytes::RawOsString;

let string = "foobar".to_owned();
let mut raw = RawOsString::new(string.clone());
raw.shrink_to_fit();
assert_eq!(string, raw);

pub fn split_off(&mut self, at: usize) -> Self

Equivalent to String::split_off.

Panics

Panics if the index is not a valid boundary.

Examples

use os_str_bytes::RawOsString;

let mut raw = RawOsString::new("foobar".to_owned());
assert_eq!("bar", raw.split_off(3));
assert_eq!("foo", raw);

pub fn truncate(&mut self, new_len: usize)

Equivalent to String::truncate.

Panics

Panics if the index is not a valid boundary.

Examples

use os_str_bytes::RawOsString;

let mut raw = RawOsString::new("foobar".to_owned());
raw.truncate(3);
assert_eq!("foo", raw);