Struct os_str_bytes::RawOsString

pub struct RawOsString(_);

A container for the byte strings converted by OsStringBytes.

For more information, see RawOsStr.

Implementations

impl RawOsString

pub fn new(string: OsString) -> Self

Converts a platform-native string into a representation that can be more easily manipulated.

For more information, see RawOsStr::new.

Examples

use std::env;

use os_str_bytes::RawOsString;

let os_string = env::current_exe()?.into_os_string();
println!("{:?}", RawOsString::new(os_string));

pub fn from_string(string: String) -> Self

Wraps a string, without copying or encoding conversion.

This method is much more efficient than RawOsString::new, since the encoding used by this crate is compatible with UTF-8.

Examples

use os_str_bytes::RawOsString;

let string = "foobar".to_owned();
let raw = RawOsString::from_string(string.clone());
assert_eq!(string, raw);

pub fn into_os_string(self) -> OsString

Converts this representation back to a platform-native string.

Examples

use std::env;

use os_str_bytes::RawOsString;

let os_string = env::current_exe()?.into_os_string();
let raw = RawOsString::new(os_string.clone());
assert_eq!(os_string, raw.into_os_string());

pub fn into_raw_vec(self) -> Vec<u8>

Returns the byte string stored by this container.

The result will match what would be returned by OsStringBytes::into_raw_vec for the same string.

Examples

use std::env;

use os_str_bytes::OsStringBytes;
use os_str_bytes::RawOsString;

let os_string = env::current_exe()?.into_os_string();
let raw = RawOsString::new(os_string.clone());
assert_eq!(os_string.into_raw_vec(), raw.into_raw_vec());

pub fn into_string(self) -> Result<String, Self>

Equivalent to OsString::into_string.

Examples

use os_str_bytes::RawOsString;

let string = "foobar".to_owned();
let raw = RawOsString::from_string(string.clone());
assert_eq!(Ok(string), raw.into_string());