Trait os_str_bytes::Pattern

pub trait Pattern: Sealed { }

Allows a type to be used for searching by RawOsStr and RawOsString.

This trait is very similar to str::pattern::Pattern, but its methods are private and it is implemented for different types.